requirements.txt
anthropic>=0.25.0
Pillow>=10.0.0
numpy>=1.24.0
PyMuPDF>=1.23.0
